Revision March 2020
The matrix.f90 file was modified to add two new matmul4 & matmul5
functions.

The matrix.f90 for the profiling exercise now uses test_matmul1, ..,
test_matmul5 functions in order to be able to profile the different
algorithms even if they are inlined.

All the Makefiles were modified for easier generation of different
optimisation executables and listing (report of optimisation) files.